requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --frozen --output-file=requirements.txt --no-dev --no-hashes
annotated-types==0.7.0
anyio==4.7.0
arclet-alconna==1.8.35
arclet-alconna-tools==0.7.10
certifi==2024.12.14
colorama==0.4.6 ; sys_platform == 'win32'
exceptiongroup==1.2.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
importlib-metadata==8.5.0
loguru==0.7.3
multidict==6.1.0
nepattern==0.7.7
nonebot-plugin-alconna==0.54.1
nonebot-plugin-waiter==0.8.0
nonebot2==2.4.1
pillow==11.0.0
propcache==0.2.1
pydantic==2.10.4
pydantic-core==2.27.2
pygtrie==2.5.0
python-dotenv==1.0.1
sniffio==1.3.1
tarina==0.6.8
tomli==2.2.1 ; python_full_version < '3.11'
typing-extensions==4.12.2
win32-setctime==1.2.0 ; sys_platform == 'win32'
yarl==1.18.3
zipp==3.21.0
